[#72] Add Data Access Objects cop
This cop checks for working with data persistence objects in controller
and suggest to move it to Data Access Objects to separate the data
access parts from the application logic.

It's important to raise violations on the lines related to this access
objects because otherwise it will be dropped by git filter.